+++ ExchangeSharp/API/Exchanges/Coinbase/Models/Types/ChannelType.cs
@@ -10,19 +10,28 @@ The above copyright notice and this permission notice shall be included in all c
 THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-namespace ExchangeSharp.CoinbaseModels
+namespace ExchangeSharp.Coinbase
 {
-    using System.Collections.Generic;
+    using System.Runtime.Serialization;
 
-    using Newtonsoft.Json;
-
-    public class Snapshot : BaseMessage
+    public enum ChannelType
     {
-        [JsonProperty("product_id")]
-        public string ProductId { get; set; }
+        [EnumMember(Value = "full")]
+        Full,
 
-        public List<decimal[]> Bids { get; set; }
+        [EnumMember(Value = "heartbeat")]
+        Heartbeat,
 
-        public List<decimal[]> Asks { get; set; }
+        [EnumMember(Value = "level2")]
+        Level2,
+
+        [EnumMember(Value = "matches")]
+        Matches,
+
+        [EnumMember(Value = "ticker")]
+        Ticker,
+
+        [EnumMember(Value = "user")]
+        User
     }
 }
