+++ ExchangeSharp/API/Exchanges/Coinbase/Models/Types/ResponseType.cs
@@ -10,28 +10,39 @@ The above copyright notice and this permission notice shall be included in all c
 THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-namespace ExchangeSharp.CoinbaseModels
+namespace ExchangeSharp.Coinbase
 {
     using System.Runtime.Serialization;
 
-    public enum ChannelType
+    public enum ResponseType
     {
-        [EnumMember(Value = "full")]
-        Full,
+        Unknown = 0,
 
-        [EnumMember(Value = "heartbeat")]
+        Subscriptions,
+
         Heartbeat,
 
-        [EnumMember(Value = "level2")]
-        Level2,
-
-        [EnumMember(Value = "matches")]
-        Matches,
-
-        [EnumMember(Value = "ticker")]
         Ticker,
 
-        [EnumMember(Value = "user")]
-        User
+        Snapshot,
+
+        L2Update,
+
+        Received,
+
+        Open,
+
+        Done,
+
+        Match,
+
+        [EnumMember(Value = "last_match")]
+        LastMatch,
+
+        Change,
+
+        Activate,
+
+        Error
     }
 }
